refactor: extract tab drag-drop logic into dedicated module
Move setupTabDrag, clearTabShifts, and updateTabDropTarget from
tab-manager.js (1006 lines) into new src/utils/tab-drag.js (154 lines).

tab-manager.js reduced from 1006 to 879 lines (-127 lines, -12.6%).
Drag state is now local per drag session instead of leaked on the
TabManager instance.
